- refactoring product list
